Move forEach to BlockingObservable
- was missed in previous refactoring
